pyproject.toml
[tool.poetry]
name = "manim"
version = "0.1.1"
description = "Animation engine for explanatory math videos."
authors = ["The Manim Community Developers","3b1b <grant@3blue1brown.com>"]
license="MIT"
readme="README.md"
repository="https://github.com/manimcommunity/manim"
documentation="https://docs.manim.community/"
classifiers= [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Graphics",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Natural Language :: English",
    ]
exclude = ["scripts/","logo/","readme-assets/"]
packages = [
    { include = "manim" },
]

[tool.poetry.dependencies]
python = "^3.6"
colour = "*"
numpy = "*"
Pillow = "*"
progressbar = "*"
scipy = "*"
tqdm = "*"
pydub = "*"
pygments = "*"
rich = "^6.0"
pycairo = "^1.19"
pangocffi = "^0.8.0"
pangocairocffi = "^0.4.0"
cairocffi = "^1.1.0"
grpcio =  { version = "*", optional = true }
grpcio-tools = { version = "*", optional = true }
watchdog = { version = "*", optional = true }

[tool.poetry.extras]
js_renderer = ["grpcio","grpcio-tools","watchdog"]

[tool.poetry.dev-dependencies]
pytest = "^6.0"
pylint = "*"
guzzle_sphinx_theme = "*"
recommonmark = "*"
matplotlib = "^3.3.2"

[tool.poetry.dev-dependencies.black]
version = "^20.8b1"
allow-prereleases = false
python = "^3.6"
markers = "platform_python_implementation == 'CPython'"

[tool.pytest.ini_options]
markers = "slow: Mark the test as slow. Can be skipped with --skip_slow"

[tool.poetry.plugins]
[tool.poetry.plugins."console_scripts"]
"manim" = "manim.__main__:main"
"manimce" = "manim.__main__:main"

[build-system]
requires = ["setuptools","poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
